move redirect to layout root
